add rating view to GameOver view and change ID to uuidv4()
