+++ music-quiz/package.json
@@ -12,6 +12,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
+    "react-rating-stars-component": "^2.2.0",
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
     "uuid": "^9.0.0",
