+++ music-quiz/src/models/Model.js
@@ -8,11 +8,6 @@ class Model {
     this.email = null;
     this.username = null;
     this.playlists = [];
-    this.settings = {
-      volume: 0.5,
-      pitch: 1,
-      speed: 1.1,
-    };
     this.pending = [];
     this.friends = [];
     this.initialLoginAttemptComplete = false;
@@ -97,17 +92,6 @@ class Model {
       playlist,
     ];
     this.notifyObservers({ editPlaylist: playlist });
-  }
-
-  getUniquePlaylistID() {
-    // Find the next largest ID that is unique.
-    // Another solution: `this.playlists.length`, however, that will not work if we are allowed to remove playlists.
-    return (
-      this.playlists.reduce(
-        (currentMax, playlist) => Math.max(currentMax, playlist.id),
-        0
-      ) + 1
-    );
   }
 
   newPendingRequest(searchUserData, requestType) {
@@ -209,13 +193,6 @@ class Model {
   clearPlaylist() {
     if (this.playlists === []) return;
     this.playlists = [];
-  }
-
-  setVolume(volume) {
-    // Set the preferred volume of the user
-    if (this.settings.volume === volume) return;
-    this.settings.volume = volume;
-    this.notifyObservers({ volume: this.settings.volume });
   }
 }
